KVM: VFIO: resolve redeclaration error on x86/_64
Compilation for x86 /x86_64 resulted in the following error:

virt/kvm/vfio.c:238:30: error:
 static declaration of 'kvm_vfio_ops' follows non-static declaration
   static struct kvm_device_ops kvm_vfio_ops
 include/linux/kvm_host.h:1057:30:
 note: previous declaration of 'kvm_vfio_ops' was here
   extern struct kvm_device_ops kvm_vfio_ops;

In the kernel.org linux-stable tree,
commit 80ce163 KVM: VFIO: register kvm_device_ops dynamically
by Will Deacon includes the removal of the declaration from
include/linux/kvm_host.h, but that portion of the change does not
seem to have made it into the 3.14 tree - while the rest of that
commit does seem to have been backported.  This patch resolves
that omission to fix the x86 compilation problems.

Acked-by: Christoffer Dall <christoffer.dall@linaro.org>
Signed-off-by: Gary S. Robertson <gary.robertson@linaro.org>
Signed-off-by: Kevin Hilman <khilman@linaro.org>