amvideocap: don't spam kernel log
When using amvideocap the kernel log is spammed each time a frame is captured.
I've commented all pr_info messages that are useful only for debugging purposes. (They can be removed or transformed into code comments)
A use case where this issue occurs is when hyperion or hyperion.ng is running.